requirements.txt
fastapi
gitpython
pydantic
uvicorn